Rename topeftenv to be more generic
